RavenDB-20652 Removing old compound sorting code / test adjust
