+++ src/Corax/Queries/SortingMatches/Comparers/SortingMatch.SpatialDescendingMatchComparer.cs
@@ -22,15 +22,15 @@ unsafe partial struct LegacySortingMatch
         private readonly SpatialUnits _units;
         
         public IPoint Point => _point;
-
+    
         public double Round => _round;
-
+    
         public SpatialUnits Units => _units;
-
+    
         public FieldMetadata Field => _field;
         
         public MatchCompareFieldType FieldType => _fieldType;
-
+    
         public SpatialDescendingMatchComparer(IndexSearcher searcher, in OrderMetadata metadata)
         {
             _searcher = searcher;
@@ -49,10 +49,10 @@ unsafe partial struct LegacySortingMatch
             {
                 var readerX = comparer._searcher.GetEntryReaderFor(x);
                 var readX = readerX.GetFieldReaderFor(comparer._field).Read(out (double lat, double lon) resultX);
-
+    
                 var readerY = comparer._searcher.GetEntryReaderFor(y);
                 var readY = readerY.GetFieldReaderFor(comparer._field).Read(out (double lat, double lon) resultY);
-
+    
                 if (readX && readY)
                 {
                     var readerXDistance = SpatialUtils.GetGeoDistance(in resultX, comparer);
@@ -62,10 +62,10 @@ unsafe partial struct LegacySortingMatch
                 }
                 else if (readX)
                     return -1;
-
+    
                 return 1;
             }
-
+    
             _compareFunc = _fieldType switch
             {
                 MatchCompareFieldType.Sequence => &ThrowOnWrongEntryFieldType,
@@ -75,13 +75,13 @@ unsafe partial struct LegacySortingMatch
                 var type => throw new NotSupportedException($"Currently, we do not support sorting by {type}.")
             };
         }
-
+    
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public int CompareNumerical<T>(T sx, T sy) where T : unmanaged, INumber<T>
         {
             return -BasicComparers.CompareAscending(sx, sy);
         }
-
+    
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public int CompareSequence(ReadOnlySpan<byte> sx, ReadOnlySpan<byte> sy)
         {
